refactor(data): 👷 texts
