+++ components/pages/home/expertise/ExpertiseSection.tsx
@@ -13,7 +13,7 @@ const ExpertiseSection = ({
   listItems = [],
 }: ExpertiseSectionProps) => {
   return (
-    <div className="relative flex w-full flex-col rounded-lg border border-gray-200 bg-white px-4 py-6 transition-shadow duration-300 ease-in-out hover:border-gray-300 hover:shadow-md">
+    <div className="relative flex w-full flex-col rounded-lg border border-gray-200 bg-white px-4 py-6 transition-shadow duration-300 ease-in-out select-none hover:border-gray-300 hover:shadow-md">
       <Heading as="h3" textColor="text-violet-600 ml-2">
         <span
           role="img"
