+++ components/pages/who-i-am/WhoIAmIntroduction.tsx
@@ -1,7 +1,7 @@
 import Highlight from '@/components/shared/Highlight'
 import Paragraph from '@/components/shared/Paragraph'
 
-import { COMMON_VALUES, INTRODUCTION, SHARED } from '@/localization/english'
+import { COMMON_VALUES, INTRODUCTION } from '@/localization/english'
 
 import { DATA_TEST_IDS } from '@/__tests__/playwright/lib/utils/constants/ids/dataTestIds'
 
@@ -11,11 +11,10 @@ const WhoIAmIntroduction = () => {
       <Paragraph dataTestId={DATA_TEST_IDS.page.whoIAm.introduction}>
         <span>{INTRODUCTION.whoIAm.text.introduction1}</span>
         <Highlight text={COMMON_VALUES.nameDanielKrsiak} />
-        {','}
-        <span>{INTRODUCTION.whoIAm.text.introduction2}</span>{' '}
-        <span className="font-bold">{INTRODUCTION.whoIAm.text.developer}</span> {SHARED.and}
+        {'.'} <span>{INTRODUCTION.whoIAm.text.introduction2}</span>{' '}
+        <span className="ml-1">{INTRODUCTION.whoIAm.text.freeTime}</span>
         <span className="ml-1">
-          <Highlight text={INTRODUCTION.whoIAm.text.freeTime} />
+          <Highlight text={INTRODUCTION.whoIAm.text.loveTravel} />
           {'.'}
         </span>
       </Paragraph>
