+++ lib/utils/typeDefinitions/props/pages/resume/resume.ts
@@ -1,16 +1,10 @@
-import { CareerPathStep, ResumePreviewCard } from '@/lib/utils/typeDefinitions/interfaces'
+import { ResumePreviewCard } from '@/lib/utils/typeDefinitions/interfaces'
 
 // Photo Props
 export type PhotoProps = {
   isPhotoSmall: boolean
   isOpenToWork?: boolean
   showCaption: boolean
-}
-
-// Career Path Job Step Props
-export type CareerPathJobStepProps = {
-  step: CareerPathStep
-  index: number
 }
 
 // Resume Preview Card
